requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
streamlit>=1.28.0
pandas>=2.0.0
pytz>=2023.3
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
